fix: 별개의 컴포넌트 락 기능 재구현
