+++ src/main/java/com/example/chatserverchat/domain/service/ChatGraphqlServiceImpl.java
@@ -3,7 +3,6 @@ package com.example.chatserverchat.domain.service;
 import com.example.chatserverchat.domain.dto.GraphqlDTO;
 import com.example.chatserverchat.domain.entity.ChatRoom;
 import com.example.chatserverchat.domain.repository.ChatRoomRepository;
-import com.example.chatserverchat.global.facade.DistributedLockFacade;
 import jakarta.ws.rs.NotFoundException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -15,7 +14,6 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.server.ResponseStatusException;
 
 import static org.apache.http.HttpStatus.SC_FORBIDDEN;
-import static org.apache.http.HttpStatus.SC_LOCKED;
 
 @Slf4j
 @Service
@@ -24,7 +22,6 @@ import static org.apache.http.HttpStatus.SC_LOCKED;
 public class ChatGraphqlServiceImpl implements ChatGraphqlService {
 
     private final ChatRoomRepository chatRoomRepository;
-    private final DistributedLockFacade distributedLock;
 
     @Override
     @Transactional(readOnly = true)
@@ -46,36 +43,17 @@ public class ChatGraphqlServiceImpl implements ChatGraphqlService {
 
     @Override
     public GraphqlDTO incrementPersonnel(String id) {
-        return getGraphqlDTO(id);
-    }
-
-    @Override
-    public GraphqlDTO incrementPersonnelWithLock(String id) {
-        String lockKey = "LOCK_" + id;
-        boolean isLocked = distributedLock.tryLock(lockKey, 5, 2);
-
-        if (!isLocked) {
-            throw new ResponseStatusException(
-                    HttpStatusCode.valueOf(SC_LOCKED), "락 획득 실패");
-        }
-
-        try {
-            return getGraphqlDTO(id);
-        } finally {
-            distributedLock.unlock(lockKey);
-        }
-    }
-
-    private GraphqlDTO getGraphqlDTO(String id) {
         ChatRoom chatRoom =
                 chatRoomRepository.findById(Long.parseLong(id)).orElseThrow(
                         () -> new IllegalArgumentException("해당하는 채팅방이 조회되지 않음")
                 );
-
         log.info("그래프큐엘 찾았냐(현재 인원 / 최대인원): {} / {}",
                 chatRoom.getPersonnel(),
                 chatRoom.getMaxPersonnel());
 
+        /**
+         * 락 구현은 이 로직 내에서 이뤄져야 할듯? 분산락 써보자
+         */
         if (chatRoom.getMaxPersonnel() <= chatRoom.getPersonnel()) {
             throw new ResponseStatusException(
                     HttpStatusCode.valueOf(SC_FORBIDDEN), "이미 정원이 초과된 채팅방");
@@ -92,6 +70,8 @@ public class ChatGraphqlServiceImpl implements ChatGraphqlService {
                 chatRoom.getCreatedAt().toString()
         );
     }
+
+
 
     @Override
     public GraphqlDTO decrementPersonnel(String id) {
